+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `libcontainer/configs.Thawed`
     - `libcontainer/configs.NewWeightDevice`
     - `libcontainer/configs.NewThrottleDevice`
+  - `libcontainer/configs.HookList.RunHooks` (#5141)
 
 ## [1.4.0] - 2025-11-27